requirements.txt
astor==0.8.1
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.0
click==7.1.2
decorator==5.1.1
future==0.18.2
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlepaddle-gpu @ file:///home/shier/Downloads/Paddle/build/python/dist/paddlepaddle_gpu-2.3.2-cp38-cp38-linux_x86_64.whl
pefile==2022.5.30
Pillow==9.2.0
protobuf==3.20.0
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.19.2
QPT==1.0b4.dev6
qt-material==2.12
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
QtPy==2.2.0
requests==2.28.1
shiboken2==5.15.2.1
six==1.16.0
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.11
wget==3.2
